Update README.md
This is to add a dependency dbus-python to the installation in Fedora, and to add instructions to install it in RHEL 7. On RHEL 7, it is not recommended to add a kernel parameter by editing /boot/config-$(uname -r) directly. Suggestions on how to correctly set CONFIG_DEVMEM=y is welcome.