openrc fix
This maintains consistency with the install.sh script by using the virtualenv that the script sets up. The original version was using a different interpreter and complained about missing imports. Additionally, the new shebang now points to the correct openrc-run executable. These were the only changes needed for this script to work on Gentoo.