Update README.md
documenting that this is working for me. I've commented on an issue the basis for my experience that it's working.

of note: I didn't try undervolting, as I know that isn't supported on my platform. 

also: if you have any steps you'd like me to take to confirm this is working before merging this request, let me know.